requirements.txt
asgiref==3.11.0
charset-normalizer==3.4.4
Django==6.0.1
djangorestframework==3.16.1
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==12.1.0
sqlparse==0.5.5
tzdata==2025.3
